T316: drop PoSConsensusModule
